user/sanzhar_u/agentic_data_pipeline/requirements.txt
langgraph>=1.0.6
langgraph-prebuilt>=1.0.6
langgraph-checkpoint>=4.0.0

langchain>=0.3.0
langchain-core>=0.3.64
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0

langgraph-cli[inmem]>=0.4.11
langgraph-api>=0.6.35
langgraph-runtime-inmem>=0.22.0

pydantic>=2.7.4
orjson>=3.9.7
httpx>=0.25.0
tenacity>=8.2.3
cloudpickle>=3.0.0
structlog>=24.1.0

uvicorn>=0.26.0
starlette>=0.38.6
sse-starlette>=2.1.0
watchfiles>=0.13

selenium>=4.18.0
requests>=2.31.0
pillow>=10.0.0

webdriver-manager>=4.0.1

python-dotenv>=1.0.0
